process-coco/requirements.txt
cycler==0.10.0
Cython==0.29.15
decorator==4.4.2
h5py==2.10.0
imageio==2.8.0
kiwisolver==1.1.0
matplotlib==3.2.0
networkx==2.4
numexpr==2.7.1
numpy==1.18.1
pandas==1.0.1
Pillow==7.0.0
pycocotools==2.0.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
scikit-image==0.16.2
scipy==1.4.1
six==1.14.0
tables==3.6.1
